data-collector-local/requirements.txt
paho-mqtt>=1.6.1
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose>=3.3.0
passlib>=1.7.4
